fix: erlang-ls issues. Remember to add deps manually in the future
This is needed because if the `_build/default/lib/hb` directory is found by erlang-ls jumping to a reference will result in moving the code editor to the post-build phase file. You may end up editing this file and losing your work. Unfortunately, erlang-ls does not offer a way to exclude directories in wildcard matches, which would obviously be nicer.