impr: `scheduler@1.0/location` key
Improves the interface for registering and retrieving scheduler-location records, including allowing nodes to notify others when their scheduler location is updated.

Additionally, the `dev_hook` interface has been updated to support requesting that the node sign requests before performing the hook execution, as well as the ability to have the result of the hook be discarded if preferred. Together, these two options make it significantly easier to integrate calls to non-hook-aware devices into hook calls.