feat: Determine node config encoding from file extension
Allows a node operator to specify the codec that should be used to deserialize their configuration file via its file extension. The first preloaded device whose name matches the file extension as a prefix is utilized. For example, to load a JSON file with the `~json@1.0` device as the config, simply ensure that its file extension ends with `.json`.

As a reminder, you can choose the file to utilize as your node configuration using the `HB_CONFIG=[path/to/your/file].[codec]` environment variable.