fix: a verification issue with ans104
1. When converting from a TABM to a #tx when there is no commitment, we sort the #tx tags alphabetically.
2. However when generating the committed keys the base fields used to be placed first.

This can cause a verification failure when `target` is set to something that is not a valid #tx.target.
In that case it gets moved into tags, and then the order can change between steps 1 and 2.
This commit attempts to fix by rearranging the order so that field_keys is last instead of first
(i.e. the tag ordering will be preserved in the event there's a tag named target)